(Issue #10) Replace target flag with prod flag
The target flag was misleading as it is commonly used as a flag to
denote the output directory for where to place rendered files.

Instead of having a target flag the flag has been replaced with a
`--prod` flag. Setting this flag will set the target to "prod" and
include github pages links.

Omitting the flag replaces the old behavior of `--target dev` so that by
default the output in the `docs` directory is linked properly without
Github pages format.

This commit addresses Issue #10.